# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Crunchy API",
        "description": "This is a sample of Crunchy API",
        "contact": {
            "name": "SmallCorridor",
            "email": "smallcorridor@gmail.com"
        },
        "version": "1.0"
    },
    "host": "crunchy-api.dev",
    "basePath": "",
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/api/v1/category": {
            "get": {
                "tags": [
                    "Category"
                ],
                "summary": "Retrieves the collection of Category resources.",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Example = Bearer(space)'your_token'",
                        "required": true,
                        "type": "string",
                        "default": "Bearer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Categories collection.",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/category"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized action."
                    }
                }
            },
            "post": {
                "tags": [
                    "Category"
                ],
                "summary": "Add new Category.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Example = Bearer(space)'your_token'",
                        "required": true,
                        "type": "string",
                        "default": "Bearer"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Category object that needs to be added to the database",
                        "required": true,
                        "schema": {
                            "properties": {
                                "name": {
                                    "type": "string"
                                },
                                "icon": {
                                    "type": "string"
                                },
                                "category": {
                                    "type": "boolean"
                                }
                            }
                        },
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "new Category has successfully added.",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/category"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized action."
                    }
                }
            }
        },
        "/api/v1/category/{id}": {
            "get": {
                "tags": [
                    "Category"
                ],
                "summary": "Find Category by ID.",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Example = Bearer(space)'your_token'",
                        "required": true,
                        "type": "string",
                        "default": "Bearer"
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Please enter the categoryId",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "This is the data that you search.",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/category"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid ID."
                    },
                    "401": {
                        "description": "Unauthorized action."
                    },
                    "404": {
                        "description": "Category not found."
                    }
                }
            },
            "put": {
                "tags": [
                    "Category"
                ],
                "summary": "Update the Category resource.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Example = Bearer(space)'your_token'",
                        "required": true,
                        "type": "string",
                        "default": "Bearer"
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Please enter the categoryId",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Category object that needs to be added to the database",
                        "required": true,
                        "schema": {
                            "properties": {
                                "name": {
                                    "type": "string"
                                },
                                "icon": {
                                    "type": "string"
                                },
                                "category": {
                                    "type": "boolean"
                                }
                            }
                        },
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "new Category has successfully updated.",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/category"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid ID."
                    },
                    "401": {
                        "description": "Unauthorized action."
                    },
                    "404": {
                        "description": "Category not found."
                    }
                }
            },
            "delete": {
                "tags": [
                    "Category"
                ],
                "summary": "Remove the Category resource.",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Example = Bearer(space)'your_token'",
                        "required": true,
                        "type": "string",
                        "default": "Bearer"
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Please enter the categoryId",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Category resource deleted."
                    },
                    "401": {
                        "description": "Unauthorized action."
                    },
                    "404": {
                        "description": "Category not found."
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "category": {
            "properties": {
                "id": {
                    "type": "integer",
                    "format": "int32"
                },
                "name": {
                    "type": "string"
                },
                "icon": {
                    "type": "string"
                },
                "category": {
                    "type": "boolean",
                    "default": true
                },
                "create_at": {
                    "type": "string"
                },
                "update_at": {
                    "type": "string"
                }
            }
        },
...
    }
}